Fix -fsanitize=undefined,unsigned-integer-overflow warnings found when running autotest/gcore
git-svn-id: https://svn.osgeo.org/gdal/trunk@39165 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/gt_jpeg_copy.cpp
@@ -702,7 +702,7 @@ static CPLErr GTIFF_CopyBlockFromJPEG( GTIFF_CopyBlockFromJPEGArgs* psArgs )
                                                         nSrcHeightInBlocks )
             {
                 const int nYBlocks =
-                    nSrcHeightInBlocks - (dst_blk_y + y_crop_blocks);
+                    (int)nSrcHeightInBlocks - (int)(dst_blk_y + y_crop_blocks);
                 if( nYBlocks > 0 )
                 {
                     JBLOCKARRAY src_buffer =
